Added role based registation and permissions

+++ src/main/java/com/rauln/CarKet/services/AdvertisementService.java
@@ -6,6 +6,7 @@ import com.rauln.CarKet.model.Car;
 import com.rauln.CarKet.model.User;
 import com.rauln.CarKet.repositories.AdvertisementRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -22,14 +23,22 @@ public class AdvertisementService {
     private final CarService carService;
     private final UserService userService;
 
-    public Advertisement save_ad(Advertisement advertisement){
+    public Advertisement saveAd(Advertisement advertisement){
         return advertisementRepository.save(advertisement);
     }
-    public void delete_ad(Long id){
-        advertisementRepository.deleteById(id);
+    public void deleteAdSecured(Long id, String userEmail, Boolean isAdmin){
+        Advertisement advertisement = advertisementRepository.findAdvertisementById(id);
+        if(advertisement.getUser().getEmail().equals(userEmail) || isAdmin) {
+            advertisementRepository.deleteById(id);
+        }else {
+            throw new AccessDeniedException("You can only delete your own ads.");
+        }
     }
     public List<Advertisement> loadAllAds(){
         return advertisementRepository.findAll();
+    }
+    public Advertisement loadAdById(Long id){
+        return advertisementRepository.findAdvertisementById(id);
     }
     public List<Advertisement> loadAdsByUser(Long user_id){
         return advertisementRepository.findAdvertisementsByUserId(user_id);
@@ -59,10 +68,10 @@ public class AdvertisementService {
             }
             Path filePath = uploadPath.resolve(fileName);
             Files.copy(image.getInputStream(), filePath);
-            advertisement.setImages(List.of("/uploads/" +fileName));
+            advertisement.setImages(List.of("/uploads/" + fileName));
         }
 
-        return save_ad(advertisement);
+        return saveAd(advertisement);
     }
 
 }
